requirements.txt
pyspark>=3.4.0
delta-spark>=2.4.0
pgeocode>=0.4.0
scipy>=1.10.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
